Fixing client id, so that never is negative

+++ Server/ConnectionHandler.java
@@ -64,22 +64,22 @@ public class ConnectionHandler implements Runnable {
         switch (response.getMessageType()) {
             case SIGNIN:
                 body = response.getBody().split(" ");
-                System.out.println("REQUEST ID: " + response.getSenderId().intValue());
+                System.out.println("REQUEST ID: " + response.getSenderId());
                 if (server.isResponsibleFor(response.getSenderId())) {
                     server.saveConnection(this.sslSocket, response.getSenderId());
                     return server.loginUser(body[0], body[1]);
                 } else {
-                    System.out.println("REDIRECTING ID: " + response.getSenderId().intValue());
+                    System.out.println("REDIRECTING ID: " + response.getSenderId());
                     server.redirect(response);
                 }
             case SIGNUP:
                 body = response.getBody().split(" ");
-                System.out.println("REQUEST ID: " + response.getSenderId().intValue());
+                System.out.println("REQUEST ID: " + response.getSenderId());
                 if (server.isResponsibleFor(response.getSenderId())) {
                     server.saveConnection(this.sslSocket, response.getSenderId());
                     return server.addUser(body[0], body[1]);
                 } else {
-                    System.out.println("REDIRECTING ID: " + response.getSenderId().intValue());
+                    System.out.println("REDIRECTING ID: " + response.getSenderId());
                     server.redirect(response);
                 }
             case CREATE_CHAT:
